Refactor MyProfile to Containcer/View Components
